BDS-1175: test

+++ yupana-hbase/src/main/scala/org/yupana/hbase/HBaseUtils.scala
@@ -116,7 +116,7 @@ object HBaseUtils extends StrictLogging {
     val startKey = List(rangeStartKey, Some(fromTimeKey), startRowKey).flatten
       .max(Ordering.comparatorToOrdering(Bytes.BYTES_COMPARATOR))
 
-    val inclusiveEndRowKey = endRowKey.filter(_.nonEmpty).map(a => a :+ 0)
+    val inclusiveEndRowKey = endRowKey.filter(_.nonEmpty).map(a => a :+ 0.toByte)
     val stopKey = List(rangeStopKey, Some(toTimeKey), inclusiveEndRowKey).flatten
       .min(Ordering.comparatorToOrdering(Bytes.BYTES_COMPARATOR))
